feat(*): add Markdown Editor example
