+++ package.json
@@ -4,15 +4,20 @@
   "description": "Starter React app for using AP web components",
   "main": "index.js",
   "dependencies": {
-    "@accordproject/ui-concerto": "^0.93.1",
-    "@accordproject/ui-markdown-editor": "^0.93.1",
+    "@accordproject/markdown-slate": "0.11.3",
+    "@accordproject/ui-concerto": "0.93.2-20200625014131",
+    "@accordproject/ui-markdown-editor": "0.93.2-20200625014131",
     "@babel/runtime": "^7.10.3",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-monaco-editor": "^0.36.0",
     "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.88.2"
+    "semantic-ui-react": "^0.88.2",
+    "slate": "^0.58.3",
+    "slate-history": "^0.58.3",
+    "slate-react": "^0.58.3",
+    "styled-components": "^5.1.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.10.3",
