+++ src/App.css
@@ -18,10 +18,6 @@ limitations under the License.
   height: 100vh;
 }
 
-/* .ui.form textarea:not([rows]) {
-  height: 100%;
-  max-height: 100%;
-} */
 .App-logo {
   animation: App-logo-spin infinite 20s linear;
   height: 80px;
@@ -46,4 +42,3 @@ limitations under the License.
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
-
